Replace all use of FIX2* with NUM2*.
This is slightly slower for converting values to something in C, the
upside is that it doesn't produce an error if the number isn't a Fixnum.
This will still produce errors if you try to write a signed integer that
exceeds the maximum (or minimum, I suppose -- that seems to be a less
common problem) possible value of an int, though.