Alias malloc as __malloc__.
Reason for this is that someone might want to override malloc and
although that's a bad idea and should never be done, it is possible and
probably relatively safe within the confines of only what snow-data
does with malloc.

I'll convert all snow-data calls to malloc to __malloc__ to sort of
guard them against overrides where possible.