Rename encoding_for_members to decode_member_info, add encode_member_info.
I figure the process should be reversible, to an extent. It's not really
possible to recreate the exact same encoding string because bits are
optional, but it is possible to create an equivalent encoding string, so
that seems like a good enough plan.

As such, CStruct::new no longer does the decoding, it's merely calling
decode_member_info to decode and handing that off to build_struct_type.