Add alloca for stack allocations.
As mentioned in the rdoc comments, this is disabled unless explicitly
enabled with --allow-alloca.

Essentially, this just lets you create a chunk of memory on the stack,
do some stuff with it, and then dispose of it just as quickly. It's not
great for security and probably not very useful for performance seeing
as you're already using ruby and so any performance benefit that you
might get is already negated by the checks I do to ensure everything is
kosher. Still, who knows, it might be useful for allocating something
and changing it many times (maybe to copy it to heap, who knows).