Add CStruct::Builder class.
This just acts as a way of building member info arrays by calling
methods rather than building the arrays yourself or providing an
encoding string. It's arguably easier to extend in the future than the
encoding string stuff, and you'll get errors if you do something wrong.

Future member declaration arguments will be added as keyword arguments,
similar to the align keyword, thereby ensuring it's explicit and obvious
what that part of the declaration does.