Replace #intern with #to_sym.
Discussed this on IRC with some folks, but the general consensus is that
to_sym is more obvious than intern in its function (they're the same,
FYI). intern suffers from being named after a C API function (rb_intern)
which means it makes more sense in code to me, but may not be as clear
to others.